requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
psycopg2==2.9.9
pycrypto==2.6.1
PyJWT==2.3.0
python-decouple==3.8
sqlparse==0.4.2
typing_extensions==4.12.2